move material options useEffect from context to MaterialFilter
